pyproject.toml
[tool.poetry]
name = "awss3"
version = "0.1.0"
description = ""
authors = ["JexPy <>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.26.93"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"